rename: NumberGenerator 패키지 이동
