+++ src/main/java/racingcar/model/RacingCars.java
@@ -1,7 +1,5 @@
 package racingcar.model;
 
-import racingcar.CarNumberGenerator;
-
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
